circularqueue.c
Implement the following Operations on Circular Queue:
Front:Get the front item from queue.
Rear: Get the last item from queue.
enQueue(value) This function is used to insert an element into the circular queue. In a circular queue, the new element is always inserted at Rear position.
1. Create a new node dynamically and insert value into it.
2. Check if front==NULL, if it is true then front = rear= (newly created node)
3. If it is false then rear=(newly created node) and rear node always contains the address of the front node.
deQueue() This function is used to delete an element from the circular queue. In a queue, the element is always deleted from front position.
1. Check whether queue is empty or not means front == NULL
2 . If it is empty then display Queue is empty. If queue is not empty then step 3
3. Check if (front-rear) if it is true then set front rear = NULL else move the front forward in queue, update address of front in rear node and return the element.